# docs/api.yaml
openapi: 3.0.3
info:
  title: LeetCode Clone API
  description: |-
    - All "Authenticated only" or "Admin only" requests are require "Authorization" header with bearer token
  version: 1.0.0
servers:
  - url: http://localhost:5000
    description: Development server
  - url: http://lcode:5000
    description: Production server run in Docker
  - url: https://lcode:5000
    description: Production server run in Docker with TLS
paths:
  /auth/register:
    post:
      tags: [ Authorization ]
      summary: Register new user
      description: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserInput'
      responses:
        200:
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        400:
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusResponse'
        500:
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusResponse'
  /auth/login:
    post:
      tags: [ Authorization ]
      summary: Login a user
      description: Login a user and return a pair of access/refresh JWT tokens
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - username
                - password
              properties:
                username:
                  type: string
                password:
                  type: string
      responses:
        200:
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tokens'
        400:
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusResponse'
        500:
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusResponse'
  /auth/refresh_tokens:
    post:
      tags: [ Authorization ]
      summary: Refresh tokens
      description: Get a new pair of JWT tokens using current refresh token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - refresh_token
              properties:
                refresh_token:
                  type: string
      responses:
        200:
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tokens'
        400:
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusResponse'
        500:
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusResponse'
  /auth/my_info:
    get:
      tags: [ Authorization ]
      summary: Get user information
      description: Authenticated users only. Get user information.
      responses:
        200:
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusResponse'
        500:
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusResponse'
  /auth/users:
    get:
      tags: [ Authorization ]
      summary: Get all users
      description: Get list of users.
      responses:
        200:
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
        400:
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusResponse'
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusResponse'
        403:
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusResponse'
        500:
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusResponse'
...
  /comments/{origin_type}/{entity_id}:
    parameters:
      - in: path
        name: origin_type
        required: true
        schema:
          type: string
          enum: [ article, problem ]
        description: Comment origin type
      - in: path
        name: entity_id
        required: true
        schema:
          type: string
          format: uuid
          example: e0c6c0e6-f3e9-4a4d-9d8c-e6e8c0e6f3e9
        description: Article ID or problem ID
    get:
      tags: [ Comments ]
      summary: Get comment threads
      description: Authenticated users only. Get comment threads with pagination for existing problem/article.
      parameters:
        - in: query
          name: sort
          schema:
            type: string
            enum: [ asc, desc ]
            default: desc
          description: Sorting by comment thread creation date order
        - in: query
          name: limit
          schema:
            type: integer
            default: 30
            description: Number of comment threads to return
        - in: query
          name: after_id
          schema:
            type: string
            format: uuid
          description: ID of the last received comment thread
      responses:
        200:
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ThreadList'
        400:
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusResponse'
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusResponse'
        500:
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusResponse'
    post:
      tags: [ Comments ]
      summary: Add comment
      description: Authenticated users only. Add comment to problem/article.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCommentInput'
      responses:
        200:
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Comment'
        400:
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusResponse'
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusResponse'
        500:
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusResponse'
  /comments/{origin_type}/{comment_id}:
    parameters:
      - in: path
        name: comment_id
        required: true
        schema:
          type: string
          format: uuid
          example: e0c6c0e6-f3e9-4a4d-9d8c-e6e8c0e6f3e9
      - in: path
        name: origin_type
        required: true
        schema:
          type: string
          enum: [ article, problem ]
        description: Comment origin type

    patch:
      tags: [ Comments ]
      summary: Update comment
      description: Authenticated users only. Update comment for existing problem/article.<br>Admins can edit all comments, users can only edit their own.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCommentInput'
      responses:
        200:
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Comment'
        400:
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusResponse'
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusResponse'
        403:
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusResponse'
        500:
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusResponse'

    delete:
      tags: [ Comments ]
      summary: Delete comment
      description: Authenticated users only. Delete comment for existing problem/article.<br>Admins can delete all comments, users can only delete their own.
      responses:
        200:
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusResponse'
        400:
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusResponse'
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusResponse'
        403:
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusResponse'
        500:
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusResponse'
components:
  schemas:
    StatusResponse:
      type: object
      required:
        - message
      properties:
        message:
          type: string
    CreateUserInput:
      type: object
      required:
        - email
        - username
        - first_name
        - last_name
        - password
      properties:
        email:
          type: string
          format: email
          maxLength: 100
        username:
          type: string
          minLength: 3
          maxLength: 50
        first_name:
          type: string
          minLength: 2
          maxLength: 50
        last_name:
          type: string
          minLength: 2
          maxLength: 50
        password:
          type: string
          minLength: 5
          maxLength: 50
...
    User:
      type: object
      required:
        - id
        - email
        - username
        - first_name
        - last_name
        - is_admin
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        username:
          type: string
        first_name:
          type: string
        last_name:
          type: string
        is_admin:
          type: boolean
    Tokens:
      type: object
      required:
        - access_token
        - access_token_exp
        - refresh_token
        - refresh_token_exp
      properties:
        access_token:
          type: string
        access_token_exp:
          type: integer
        refresh_token:
          type: string
        refresh_token_exp:
          type: integer
...
    Pagination:
      type: object
      required:
        - after_id
      properties:
        after_id:
          type: string
          format: uuid
          description: Last item ID from given page
          example: 5c7f8d0d-5a1e-4b7d-a1b9-d2b0a2c3c4d2
    UpdateCommentInput:
      type: object
      required:
        - comment_text
      properties:
        comment_text:
          type: string
          description: Comment text
          example: Ipsum dolor sit amet, consectetur adipiscing elit
    CreateCommentInput:
      type: object
      allOf:
        - $ref: '#/components/schemas/UpdateCommentInput'
      required:
        - comment_text
      properties:
        parent_id:
          type: string
          format: uuid
          description: Parent comment ID (if any)
          example: c6d0c29e-aa2d-45c5-b203-bbf9ecf41384
    Comment:
      type: object
      allOf:
        - $ref: '#/components/schemas/CreateCommentInput'
      required:
        - id
        - parent_id
        - entity_id
        - comment_text
        - created_at
        - author
      properties:
        id:
          type: string
          format: uuid
          description: Comment ID
          example: c6d0c29e-aa2d-45c5-b203-bbf9ecf41384
        created_at:
          type: integer
          format: uint64
          description: Creation time in milliseconds in standard UNIX format
          example: 1705417437
        author:
          type: object
          required:
            - user_id
            - username
            - first_name
            - last_name
          properties:
            user_id:
              type: string
              format: uuid
              description: User ID
              example: c6d0c29e-aa2d-45c5-b203-bbf9ecf41384
            username:
              type: string
              description: User username
              example: peter_parker_1337
            first_name:
              type: string
              description: User first name
              example: Peter
            last_name:
              type: string
              description: User last name
              example: Parker
    Thread:
      type: object
      required:
        - comment
        - replies
      properties:
        comment:
          $ref: '#/components/schemas/Comment'
        replies:
          type: array
          items:
            $ref: '#/components/schemas/Comment'
    ThreadList:
      type: object
      required:
        - threads
        - pagination
      properties:
        threads:
          type: array
          items:
            $ref: '#/components/schemas/Thread'
        pagination:
          $ref: '#/components/schemas/Pagination'
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - BearerAuth: [ ]